Small changes to using byte instead of UInt.

+++ MainWindow.xaml.cs
@@ -24,13 +24,13 @@ namespace Internet_Speed_Test
     public partial class MainWindow : Window
     {
         private const int HOTKEY_ID = 9000;
-        private const uint MOD_NONE = 0x0000; //[NONE]
-        private const uint MOD_ALT = 0x0001; //ALT
-        private const uint MOD_CONTROL = 0x0002; //CTRL
-        private const uint MOD_SHIFT = 0x0004; //SHIFT
-        private const uint MOD_WIN = 0x0008; //WINDOWS
+        private const byte MOD_NONE = 0x0000; //[NONE]
+        private const byte MOD_ALT = 0x0001; //ALT
+        private const byte MOD_CONTROL = 0x0002; //CTRL
+        private const byte MOD_SHIFT = 0x0004; //SHIFT
+        private const byte MOD_WIN = 0x0008; //WINDOWS
                                              //CAPS LOCK:
-        private const uint VK_CAPITAL = 0x49;
+        private const byte VK_CAPITAL = 0x49;
         [DllImport("user32.dll")]
         private static extern bool RegisterHotKey(IntPtr hWnd, int id, uint fsModifiers, uint vk);
 
@@ -59,7 +59,7 @@ namespace Internet_Speed_Test
 
         private IntPtr HwndHook(IntPtr hwnd, int msg, IntPtr wParam, IntPtr lParam, ref bool handled)
         {
-            const int WM_HOTKEY = 0x0312;
+            const short WM_HOTKEY = 0x0312;
             switch (msg)
             {
                 case WM_HOTKEY:
